publisher/requirements.txt
fastapi
uvicorn
openai
azure-servicebus